version 0.0.1

+++ src/main/java/com/speech/up/user/service/UserService.java
@@ -1,7 +1,7 @@
 package com.speech.up.user.service;
 
 import com.speech.up.common.exception.http.BadRequestException;
-import com.speech.up.oAuth.provider.JwtProvider;
+import com.speech.up.auth.provider.JwtProvider;
 import com.speech.up.user.entity.UserEntity;
 import com.speech.up.user.repository.UserRepository;
 import com.speech.up.user.service.dto.UserGetInfoDto;
@@ -30,10 +30,16 @@ public class UserService {
         return UserGetInfoDto.UserGetInfoResponseDto.getUserInfo(userEntity);
     }
 
-    // 조회?
-    // 탈퇴
-
-    public void deleteUser(String socialId) {
+    public void deleteUser(HttpServletRequest request) {
+        String authorization = request.getHeader("Authorization");
+        if(authorization == null) {
+            throw new BadRequestException("Authorization header is missing");
+        }
+        if(authorization.startsWith("Bearer ")) {
+            authorization = authorization.substring(7);
+        }
+        String socialId = jwtProvider.validate(authorization);
+        System.out.println(socialId);
         userRepository.deleteBySocialId(socialId);
     }
 }
